rename methods to match mark-all-as-read (in good cargo-cult fashion)
